Extract grid background into a separate utils file
