+++ frontend/src/metabase/dashboard/components/grid/GridLayout.jsx
@@ -1,12 +1,11 @@
 /* eslint-disable react/prop-types */
 import React, { useCallback, useMemo } from "react";
 import ReactGridLayout from "react-grid-layout";
-import _ from "underscore";
-
-import { color } from "metabase/lib/colors";
 
 import "react-grid-layout/css/styles.css";
 import "react-resizable/css/styles.css";
+
+import { generateGridBackground } from "./utils";
 
 function GridLayout({
   items,
@@ -29,29 +28,6 @@ function GridLayout({
     };
   }, [cols, gridWidth, rowHeight, margin]);
 
-  const background = useMemo(() => {
-    const XMLNS = "http://www.w3.org/2000/svg";
-    const rowHeight = cellSize.height + margin;
-    const cellStrokeColor = color("border");
-
-    const y = 0;
-    const w = cellSize.width;
-    const h = cellSize.height;
-
-    const rectangles = _(cols).times(i => {
-      const x = i * (cellSize.width + margin);
-      return `<rect stroke='${cellStrokeColor}' stroke-width='1' fill='none' x='${x}' y='${y}' width='${w}' height='${h}'/>`;
-    });
-
-    const svg = [
-      `<svg xmlns='${XMLNS}' width='${gridWidth}' height='${rowHeight}'>`,
-      ...rectangles,
-      `</svg>`,
-    ].join("");
-
-    return `url("data:image/svg+xml;utf8,${encodeURIComponent(svg)}")`;
-  }, [cellSize, gridWidth, margin, cols]);
-
   const renderItem = useCallback(
     item => {
       const itemLayout = layout.find(l => String(l.i) === String(item.id));
@@ -71,6 +47,11 @@ function GridLayout({
     }
     return (cellSize.height + margin) * lowestLayoutCellPoint;
   }, [cellSize.height, layout, margin, isEditing]);
+
+  const background = useMemo(
+    () => generateGridBackground({ cellSize, margin, cols, gridWidth }),
+    [cellSize, gridWidth, margin, cols],
+  );
 
   const style = useMemo(
     () => ({
